+++ CHANGELOG.md
@@ -37,6 +37,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MetricType::is_stateset()`
 - `MetricType::is_unknown()`
 - `AlertingRule::keep_firing_for()`
+- `impl Eq for WalReplayState`
+- `WalReplayState::is_waiting()`
+- `WalReplayState::is_in_progress()`
+- `WalReplayState::is_done()`
 
 ### Changed
 - `Alert::value()` now returns f64.